[chore] Split tests into topical files.
