+++ test/jmespath-parser.spec.js
@@ -1,0 +1,7 @@
+import { compile } from '../src';
+
+describe('parsing', () => {
+  it('should parse field node', () => {
+    expect(compile('foo')).toMatchObject({ type: 'Field', name: 'foo' });
+  });
+});
